requirements.txt
streamlit==1.32.0
newsapi-python==0.2.7
newspaper3k==0.2.8
transformers==4.38.2
torch==2.2.1
plotly==5.19.0
pandas==2.2.1
python-dotenv==1.0.1
requests==2.31.0 